+++ demisto_sdk/commands/update_release_notes/update_rn_manager.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 from typing import Optional, Tuple
 
 import git
@@ -20,7 +21,7 @@ from demisto_sdk.commands.validate.validate_manager import ValidateManager
 class UpdateReleaseNotesManager:
     def __init__(self, user_input: Optional[str] = None, update_type: Optional[str] = None,
                  pre_release: bool = False, is_all: Optional[bool] = False, text: Optional[str] = None,
-                 specific_version: Optional[str] = None, id_set_path: Optional[str] = None,
+                 specific_version: Optional[str] = None, id_set_path: Optional[Path] = None,
                  prev_ver: Optional[str] = None, is_force: bool = False, is_bc: bool = False):
         self.given_pack = user_input
         self.changed_packs_from_git: set = set()