+++ demisto_sdk/commands/format/tests/test_formatting_readme_test.py
@@ -10,6 +10,18 @@ INVALID_MD = f'{git_path()}/demisto_sdk/tests/test_files/README-invalid.md'
 INVALID_MD_IN_PACK = f'{git_path()}/demisto_sdk/tests/test_files/Packs/DummyPack2'
 
 
+def test_format_wit_update_docker_flag(mocker):
+    """
+    Check when run demisto-sdk format execute with -ud (update docker) from repo which does not have a mdx server,
+    (but has a node), that the run ends without any exception.
+    """
+    from demisto_sdk.commands.common.hook_validations.readme import \
+        ReadMeValidator
+    from demisto_sdk.commands.format.format_module import format_manager
+    mocker.patch.object(ReadMeValidator, 'are_modules_installed_for_verify', return_value=False)
+    assert format_manager(input=f'{git_path()}/Packs/TestPack', update_docker=True) == 0
+
+
 def get_new_url_from_user_assume_yes(relative_url: list) -> Optional[str]:
     """Check if new url is as expected when using assume_yes flag"""
     readme_formatter = ReadmeFormat(INVALID_MD, assume_yes=True)