+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 * Fixed an issue in the **format** command where fail when executed from environment without mdx server available.
 * Added `Added a`, `Added an` to the list of allowed changelog prefixes.
-* Added new validation to the **validate** command to check the playbook task's value references.
+* Added support for Indicator Types/Reputations in the **upload** command.
+* Fixed an issue when running from a subdirectory of a content repo failed.
+* Added a new validation to the **validate** command to check the playbook task's value references.
 
 ## 1.7.2