+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 * Added `Added a`, `Added an` to the list of allowed changelog prefixes.
 * Added support for Indicator Types/Reputations in the **upload** command.
 * Fixed an issue when running from a subdirectory of a content repo failed.
-* Changing the way we are using XSIAM servers api-keys in ***test-content*** command.
-
+* Changing the way we are using XSIAM servers api-keys in **test-content** .
+* Added a success message to **postman-codegen**.
 
 ## 1.7.2