+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Changelog
 
 ## Unreleased
+* Fixed an issue in the **format** command where fail when executed from environment without mdx server available.
+* Added `Added a`, `Added an` to the list of allowed changelog prefixes.
+* Added support for Indicator Types/Reputations in the **upload** command.
+* Fixed an issue when running from a subdirectory of a content repo failed.
+* Changing the way we are using XSIAM servers api-keys in ***test-content*** command.
+
+
+## 1.7.2
+
+* Fixed an issue in the **validate** command where incident fields were not found in mappers even when they exist
+* Added an ability to provide list of marketplace names as a param attribute to **validate** and **upload**
+* Added the file type to the error message when it is not supported.
+* Fixed an issue where `contribution_converter` incorrectly mapped _Indicator Field_ objects to the _incidentfield_ directory in contribution zip files.
+* Fixed a bug where **validate** returned error on empty inputs not used in playbooks.
+* Added the `DEMISTO_SDK_CONTENT_PATH` environment variable, implicitly used in various commands.
+* Added link to documentation for error messages regarding use cases and tags.
 
 ## 1.7.1